Allow arbitrary Redis client options to be passed via plugin config
The current version of the module only recognizes redis_server and
redis_debug, which does not permit module users to use a redis server
that requires authentication, or to tune reconnection parameters.

This patch modifies Redis.pm to pass every config key beginning with
redis_ to the Redis client constructor.

Signed-off-by: Chase Venters <chase.venters@chaseventers.com>